@@ -49,7 +49,7 @@
     "emmet-codemirror": "^1.2.5",
     "errorhandler": "^1.4.2",
     "es6-map": "~0.1.1",
-    "eslint": "~1.10.2",
+    "eslint": "~2.2.0",
     "eslint-plugin-react": "^4.1.0",
     "express": "^4.13.3",
     "express-flash": "~0.0.2",
@@ -61,7 +61,7 @@
     "gulp": "^3.9.0",
     "gulp-babel": "^6.1.1",
     "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^1.1.0",
+    "gulp-eslint": "^2.0.0",
     "gulp-jsonlint": "^1.1.0",
     "gulp-less": "^3.0.3",
     "gulp-nodemon": "^2.0.3",
@@ -74,7 +74,7 @@
     "gulp-util": "^3.0.6",
     "helmet": "^1.1.0",
     "helmet-csp": "^1.0.3",
-    "history": "^1.17.0",
+    "history": "^2.0.0",
     "jade": "^1.11.0",
     "json-loader": "~0.5.2",
     "less": "^2.5.1",
@@ -92,6 +92,7 @@
     "node-uuid": "^1.4.3",
     "nodemailer": "^2.1.0",
     "normalize-url": "^1.3.1",
+    "normalizr": "^2.0.0",
     "object.assign": "^4.0.3",
     "passport-facebook": "^2.0.0",
     "passport-github": "^1.0.0",
@@ -105,11 +106,18 @@
     "react-bootstrap": "~0.28.1",
     "react-dom": "~0.14.3",
     "react-motion": "~0.4.2",
-    "react-router": "^1.0.0",
-    "react-router-bootstrap": "https://github.com/FreeCodeCamp/react-router-bootstrap.git#freecodecamp",
+    "react-pure-render": "^1.0.2",
+    "react-redux": "^4.0.6",
+    "react-router": "^2.0.0",
+    "react-router-bootstrap": "~0.20.1",
     "react-toastr": "^2.4.0",
+    "react-router-redux": "^2.1.0",
     "react-vimeo": "~0.1.0",
+    "redux": "^3.0.5",
+    "redux-actions": "^0.9.1",
+    "redux-form": "^4.1.4",
     "request": "^2.65.0",
+    "reselect": "^2.0.2",
     "rev-del": "^1.0.5",
     "rx": "^4.0.0",
     "sanitize-html": "^1.11.1",
