@@ -92,18 +92,16 @@ window.common = (function({ $, common = { init: [] }}) {
   }
 
   function handleActionClick(e) {
-    var props = common.challengeSeed[0] ||
-      { stepIndex: [] };
+    var props = common.challengeSeed[0] || { stepIndex: [] };
 
     var $el = $(this);
     var index = +$el.attr('id');
     var propIndex = props.stepIndex.indexOf(index);
 
     if (propIndex === -1) {
-      return $el
-      .parent()
-      .find('.disabled')
-      .removeClass('disabled');
+      return $el.parent()
+        .find('.disabled')
+        .removeClass('disabled');
     }
 
     // an API action
@@ -112,30 +110,26 @@ window.common = (function({ $, common = { init: [] }}) {
     var prop = props.properties[propIndex];
     var api = props.apis[propIndex];
     if (common[prop]) {
-      return $el
-      .parent()
-      .find('.disabled')
-      .removeClass('disabled');
-    }
-    $
-    .post(api)
-    .done(function(data) {
-      // assume a boolean indicates passing
-      if (typeof data === 'boolean') {
-        return $el
-        .parent()
+      return $el.parent()
         .find('.disabled')
         .removeClass('disabled');
-      }
-      // assume api returns string when fails
-      $el
-      .parent()
-      .find('.disabled')
-      .replaceWith('<p>' + data + '</p>');
-    })
-    .fail(function() {
-      console.log('failed');
-    });
+    }
+    return $.post(api)
+      .done(function(data) {
+        // assume a boolean indicates passing
+        if (typeof data === 'boolean') {
+          return $el.parent()
+            .find('.disabled')
+            .removeClass('disabled');
+        }
+        // assume api returns string when fails
+        return $el.parent()
+          .find('.disabled')
+          .replaceWith('<p>' + data + '</p>');
+      })
+      .fail(function() {
+        console.log('failed');
+      });
   }
 
   function handleFinishClick(e) {
@@ -199,6 +193,7 @@ window.common = (function({ $, common = { init: [] }}) {
     $(nextBtnClass).click(handleNextStepClick);
     $(actionBtnClass).click(handleActionClick);
     $(finishBtnClass).click(handleFinishClick);
+    return null;
   });
 
   return common;
