@@ -1,95 +1,102 @@
 import React, { PropTypes } from 'react';
-import { contain } from 'thundercats-react';
+import { connect } from 'react-redux';
 import { Button, Col, Row } from 'react-bootstrap';
-import { History } from 'react-router';
 import Vimeo from 'react-vimeo';
-import debugFactory from 'debug';
+import { createSelector } from 'reselect';
+import debug from 'debug';
 
-const debug = debugFactory('freecc:hikes');
+import { hardGoTo } from '../../../redux/actions';
+import { toggleQuestionView } from '../redux/actions';
+import { getCurrentHike } from '../redux/selectors';
 
-export default contain(
-  {
-    actions: ['hikesActions'],
-    store: 'appStore',
-    map(state) {
-      const {
-        currentHike: {
-          dashedName,
-          description,
-          challengeSeed: [id] = [0]
-        } = {}
-      } = state.hikesApp;
+const log = debug('fcc:hikes');
 
-      return {
-        dashedName,
-        description,
-        id
-      };
-    }
-  },
-  React.createClass({
-    displayName: 'Lecture',
-    mixins: [History],
-
-    propTypes: {
-      dashedName: PropTypes.string,
-      description: PropTypes.array,
-      id: PropTypes.string,
-      hikesActions: PropTypes.object
-    },
-
-    shouldComponentUpdate(nextProps) {
-      const { props } = this;
-      return nextProps.id !== props.id;
-    },
-
-    handleError: debug,
-
-    handleFinish(hikesActions) {
-      debug('loading questions');
-      hikesActions.toggleQuestions();
-    },
-
-    renderTranscript(transcript, dashedName) {
-      return transcript.map((line, index) => (
-        <p
-          className='lead text-left'
-          key={ dashedName + index }>
-          { line }
-        </p>
-      ));
-    },
-
-    render() {
-      const {
-        id = '1',
-        description = [],
-        hikesActions
-      } = this.props;
-      const dashedName = 'foo';
-
-      return (
-        <Col xs={ 12 }>
-          <Row>
-            <Vimeo
-              onError={ this.handleError }
-              onFinish= { () => this.handleFinish(hikesActions) }
-              videoId={ id } />
-          </Row>
-          <Row>
-            <article>
-              { this.renderTranscript(description, dashedName) }
-            </article>
-            <Button
-              block={ true }
-              bsSize='large'
-              bsStyle='primary'
-              onClick={ () => this.handleFinish(hikesActions) }>
-              Take me to the Questions
-            </Button>
-          </Row>
-        </Col>
-      );
-    }
-  })
+const mapStateToProps = createSelector(
+  getCurrentHike,
+  (currentHike) => {
+    const {
+      dashedName,
+      description,
+      challengeSeed: [id] = [0]
+    } = currentHike;
+    return {
+      id,
+      dashedName,
+      description
+    };
+  }
 );
+
+export class Lecture extends React.Component {
+  static displayName = 'Lecture';
+
+  static propTypes = {
+    // actions
+    toggleQuestionView: PropTypes.func,
+    // ui
+    id: PropTypes.number,
+    description: PropTypes.array,
+    dashedName: PropTypes.string,
+    hardGoTo: PropTypes.func
+  };
+
+  componentWillMount() {
+    if (!this.props.id) {
+      this.props.hardGoTo('/map');
+    }
+  }
+
+  shouldComponentUpdate(nextProps) {
+    const { props } = this;
+    return nextProps.id !== props.id;
+  }
+
+  handleError: log;
+
+  renderTranscript(transcript, dashedName) {
+    return transcript.map((line, index) => (
+      <p
+        className='lead text-left'
+        key={ dashedName + index }
+        dangerouslySetInnerHTML={{__html: line}} />
+    ));
+  }
+
+  render() {
+    const {
+      id = '1',
+      description = [],
+      toggleQuestionView
+    } = this.props;
+
+    const dashedName = 'foo';
+
+    return (
+      <Col xs={ 12 }>
+        <Row>
+          <Vimeo
+            onError={ this.handleError }
+            onFinish= { toggleQuestionView }
+            videoId={ '' + id } />
+        </Row>
+        <Row>
+          <article>
+            { this.renderTranscript(description, dashedName) }
+          </article>
+          <Button
+            block={ true }
+            bsSize='large'
+            bsStyle='primary'
+            onClick={ toggleQuestionView }>
+            Take me to the Questions
+          </Button>
+        </Row>
+      </Col>
+    );
+  }
+}
+
+export default connect(
+  mapStateToProps,
+  { hardGoTo, toggleQuestionView }
+)(Lecture);
