@@ -14,10 +14,10 @@ export default React.createClass({
       hikes = [{}]
     } = this.props;
 
-    const vidElements = hikes.map(({ title, dashedName}) => {
+    const vidElements = hikes.map(({ title, dashedName }) => {
       return (
         <ListGroupItem key={ dashedName }>
-          <Link to={ `/hikes/${dashedName}` }>
+          <Link to={ `/videos/${dashedName}` }>
             <h3>{ title }</h3>
           </Link>
         </ListGroupItem>
