@@ -36,8 +36,9 @@ export default function addReturnToUrl() {
     ) {
       return next();
     }
-    req.session.returnTo = req.originalUrl === '/map-aside'
-    ? '/map' : req.originalUrl;
-    next();
+    req.session.returnTo = req.originalUrl === '/map-aside' ?
+      '/map' :
+      req.originalUrl;
+    return next();
   };
 }
