@@ -1,23 +1,23 @@
 import debugFactory from 'debug';
 import assign from 'object.assign';
 
-const debug = debugFactory('freecc:services:hikes');
+const debug = debugFactory('fcc:services:hikes');
 
 export default function hikesService(app) {
   const Challenge = app.models.Challenge;
 
   return {
     name: 'hikes',
-    read: (req, resource, params, config, cb) => {
+    read: (req, resource, { dashedName } = {}, config, cb) => {
       const query = {
         where: { challengeType: '6' },
         order: ['order ASC', 'suborder ASC' ]
       };
 
-      debug('params', params);
-      if (params) {
+      debug('dashedName', dashedName);
+      if (dashedName) {
         assign(query.where, {
-          dashedName: { like: params.dashedName, options: 'i' }
+          dashedName: { like: dashedName, options: 'i' }
         });
       }
       debug('query', query);
@@ -25,7 +25,7 @@ export default function hikesService(app) {
         if (err) {
           return cb(err);
         }
-        cb(null, hikes);
+        return cb(null, hikes.map(hike => hike.toJSON()));
       });
     }
   };
